Begin refactoring Redux code to match the directory/file structure in this article: https://joshuaavalon.io/create-type-safe-react-redux-store-with-typescript
