+++ myTSReduxApp/store/configureStore.ts
@@ -1,5 +1,5 @@
 import {createStore, combineReducers} from 'redux';
-import pageReducer from '../reducers/pageReducer';
+import pageReducer from './pageReducer/reducer';
 import {applyMiddleware} from 'redux';
 import thunk from 'redux-thunk';
 const rootReducer = combineReducers({pageList: pageReducer});
